Verificação numérica agora é feita usando long para evitar problemas. Alterações na classe banco.

+++ src/java/Classes/Global.java
@@ -4,7 +4,7 @@ public class Global {
     public static boolean isNumeric(String valor){
         boolean isNumeric = false;
         try {
-            int n = Integer.parseInt(valor);
+            long n = Long.parseLong(valor);
             isNumeric = true;
         }
         catch (Exception ex) {
